Add c++11 std flag in bindigs.gyp file. This is the default for node 4.0+. This allows pre-4.0/pre-iojs environments to compile with Nan 2.x Nan 2.x requires a c++11 compliant compiler.
This still doesn't allow clang to be used under a 0.12.7 environment.
However, the previous version didn't allow clang either.
A good TODO would be to figure out how to turn off the old -fno-tree-sink flag
as clang doesn't support it and node-gyp in v0.12.7 sets it by default.

The tests in the /test directory were run under the follwing environments:
 node: 4.1.1
  g++: 4.9.2
==> PASSED

 node: 4.1.1
clang: 3.6.0
==> PASSED

 node: 0.12.7
  g++: 4.9.2
==> PASSED